Refactor StoryUpload
